Move to extras and update documentation
