+++ wsidicomizer/extras/bioformats/bioformats_reader.py
@@ -12,6 +12,8 @@
 #    See the License for the specific language governing permissions and
 #    limitations under the License.
 
+"""Reader for bioformat supported files."""
+
 import os
 from contextlib import contextmanager
 from functools import cached_property, lru_cache
